Simplify fixdiv and attempt to fix automap spheres being way too large sometimes.
This didn't work, however, since points that fail don't appear to be overflowing. The point is so close to the plane that it registers as having a massive radius after dividing, but it's not overflowing during the calcs. There may still be something wrong with the clipper or something else in the pipeline. It may also be a problem in the original Descent 2, but I can't seem to cause it to happen.