Add support for the SDL "GUI" keys
a.k.a. "Command", "Meta", or "Windows Key".

Modeled after the original code for D1 and D2 for Macintosh. We have
both "LCMD" and "RCMD" in the kconfig screen, but back then it was just
one key called "APL".

The KEY_<X> constants in key.h are meaningless for SDL, but were
originally based on the DOS scancodes (which then carried over to
Windows "Virtual-Key Codes"). Therefore values for KEY_LCMD and KEY_RCMD
were chosen to correspond to "Left Windows key" and "Right Windows key"
when compiling without SDL on Windows.