Restore missing help text
The cockpit and screen resizing were not available in "VR modes", so
those keys weren't supposed to be displayed in the help text when in VR
mode. It looks like they got removed permanently by mistake.

The Descent II help menu just shows those keys all the time, so why not
do that here too.